+++ server/projects/kikbak/src/com/kikbak/dao/ReadOnlyUser2FriendDAO.java
@@ -1,6 +1,5 @@
 package com.kikbak.dao;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import com.kikbak.dto.User2friend;
@@ -9,5 +8,5 @@ public interface ReadOnlyUser2FriendDAO {
 
 	public User2friend findById(Long id);
 	public Collection<Long> listFriendsForUser(Long userId);
-	public Collection<BigInteger> listFriendsToDelete(Long userId, Collection<Long> facebookIds);
+	public Collection<Long> listFriendsToDelete(Long userId, Collection<Long> facebookIds);
 }
